+++ src/index.ts
@@ -9,7 +9,8 @@ export {
     IComplexType,
     IModelType,
     ISnapshottable,
-    IExtendedObservableMap
+    IExtendedObservableMap,
+    typecheckPublic as typecheck
 } from "./types"
 
 export * from "./core/mst-operations"
@@ -31,3 +32,5 @@ export {
     ISerializedActionCall,
     recordActions
 } from "./middlewares/on-action"
+
+export { createActionTrackingMiddleware } from "./middlewares/create-action-tracking-middleware"