+++ src/types/type-checker.ts
@@ -86,13 +86,21 @@ export function flattenTypeErrors(errors: IValidationResult[]): IValidationResul
 export function typecheck(type: IType<any, any>, value: any): void {
     // if not in dev-mode, do not even try to run typecheck. Everything is developer fault!
     if (process.env.NODE_ENV === "production") return
+    typecheckPublic(type, value)
+}
 
-    const errors = type.validate(value, [
-        {
-            path: "",
-            type
-        }
-    ])
+/**
+ * Run's the typechecker on the given type.
+ * Throws if the given value is not according the provided type specification.
+ * Use this if you need typechecks even in a production build (by default all automatic runtime type checks will be skipped in production builds)
+ *
+ * @alias typecheck
+ * @export
+ * @param {IType<any, any>} type
+ * @param {*} value
+ */
+export function typecheckPublic(type: IType<any, any>, value: any): void {
+    const errors = type.validate(value, [{ path: "", type }])
 
     if (errors.length > 0) {
         fail(